made autopilot less obtuse. made object actions menu thing not disappear immediately at time accel
git-svn-id: https://pioneer.svn.sourceforge.net/svnroot/pioneer/trunk@389 e632f14b-6550-0410-b89e-a82653faca30

+++ src/WorldView.h
@@ -55,7 +55,7 @@ private:
 	bool m_labelsOn;
 	enum CamType m_camType;
 	int m_numLights;
-	float m_showTargetActionsTimeout;
+	Uint32 m_showTargetActionsTimeout;
 };
 
 #endif /* _WORLDVIEW_H */
